app/backend/requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
